Move ApiResult to Domain
