+++ app/src/main/java/com/bottlerocketstudios/brarchitecture/ui/BaseViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.fragment.findNavController
-import com.bottlerocketstudios.brarchitecture.data.model.ApiResult
+import com.bottlerocketstudios.brarchitecture.domain.models.ApiResult
 import com.bottlerocketstudios.brarchitecture.infrastructure.coroutine.DispatcherProvider
 import com.bottlerocketstudios.brarchitecture.infrastructure.toast.Toaster
 import com.bottlerocketstudios.brarchitecture.navigation.ExternalNavigationEvent
