+++ data/src/main/java/com/bottlerocketstudios/brarchitecture/data/repository/BitbucketRepository.kt
@@ -1,19 +1,19 @@
 package com.bottlerocketstudios.brarchitecture.data.repository
 
-import com.bottlerocketstudios.brarchitecture.data.model.ApiResult
 import com.bottlerocketstudios.brarchitecture.data.model.RepoFile
 import com.bottlerocketstudios.brarchitecture.data.model.Repository
 import com.bottlerocketstudios.brarchitecture.data.model.ResponseToApiResultMapper
 import com.bottlerocketstudios.brarchitecture.data.model.Snippet
 import com.bottlerocketstudios.brarchitecture.data.model.User
 import com.bottlerocketstudios.brarchitecture.data.model.ValidCredentialModel
-import com.bottlerocketstudios.brarchitecture.data.model.alsoOnSuccess
-import com.bottlerocketstudios.brarchitecture.data.model.asSuccess
-import com.bottlerocketstudios.brarchitecture.data.model.map
-import com.bottlerocketstudios.brarchitecture.data.model.wrapExceptions
 import com.bottlerocketstudios.brarchitecture.data.network.BitbucketService
 import com.bottlerocketstudios.brarchitecture.data.network.auth.BitbucketCredentialsRepository
 import com.bottlerocketstudios.brarchitecture.data.network.auth.token.TokenAuthService
+import com.bottlerocketstudios.brarchitecture.domain.models.ApiResult
+import com.bottlerocketstudios.brarchitecture.domain.models.alsoOnSuccess
+import com.bottlerocketstudios.brarchitecture.domain.models.asSuccess
+import com.bottlerocketstudios.brarchitecture.domain.models.map
+import com.bottlerocketstudios.brarchitecture.domain.models.wrapExceptions
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import okhttp3.MediaType
@@ -89,7 +89,7 @@ internal class BitbucketRepositoryImpl(
     }
 
     override suspend fun refreshUser(): ApiResult<Unit> {
-        return wrapExceptions("refreshUser") {
+        return wrapRepoExceptions("refreshUser") {
             bitbucketService.getUser().toResult().alsoOnSuccess {
                 _user.value = it
             }.map { Unit.asSuccess() }
@@ -97,7 +97,7 @@ internal class BitbucketRepositoryImpl(
     }
 
     override suspend fun refreshMyRepos(): ApiResult<Unit> {
-        return wrapExceptions("refreshMyRepos") {
+        return wrapRepoExceptions("refreshMyRepos") {
             // TODO: Add support for handling multiple workspaces, as using the user.username might only map to the first workspace created.
             bitbucketService.getRepositories(_user.value?.username ?: "").toResult().alsoOnSuccess {
                 _repos.value = it.values.orEmpty()
@@ -106,7 +106,7 @@ internal class BitbucketRepositoryImpl(
     }
 
     override suspend fun refreshMySnippets(): ApiResult<Unit> {
-        return wrapExceptions("refreshMySnippets") {
+        return wrapRepoExceptions("refreshMySnippets") {
             bitbucketService.getSnippets().toResult().map { it.values.orEmpty().asSuccess() }.alsoOnSuccess { snippets: List<Snippet> ->
                 _snippets.value = snippets
             }.map { Unit.asSuccess() }
@@ -114,37 +114,37 @@ internal class BitbucketRepositoryImpl(
     }
 
     override suspend fun getRepositories(workspaceSlug: String): ApiResult<List<Repository>> {
-        return wrapExceptions("getRepositories") {
+        return wrapRepoExceptions("getRepositories") {
             bitbucketService.getRepositories(workspaceSlug).toResult().map { it.values.orEmpty().asSuccess() }
         }
     }
 
     override suspend fun getRepository(workspaceSlug: String, repo: String): ApiResult<Repository> {
-        return wrapExceptions("getRepository") {
+        return wrapRepoExceptions("getRepository") {
             bitbucketService.getRepository(workspaceSlug, repo).toResult()
         }
     }
 
     override suspend fun getSource(workspaceSlug: String, repo: String): ApiResult<List<RepoFile>> {
-        return wrapExceptions("getSource") {
+        return wrapRepoExceptions("getSource") {
             bitbucketService.getRepositorySource(workspaceSlug, repo).toResult().map { it.values.orEmpty().asSuccess() }
         }
     }
 
     override suspend fun getSourceFolder(workspaceSlug: String, repo: String, hash: String, path: String): ApiResult<List<RepoFile>> {
-        return wrapExceptions("getSourceFolder") {
+        return wrapRepoExceptions("getSourceFolder") {
             bitbucketService.getRepositorySourceFolder(workspaceSlug, repo, hash, path).toResult().map { it.values.orEmpty().asSuccess() }
         }
     }
 
     override suspend fun getSourceFile(workspaceSlug: String, repo: String, hash: String, path: String): ApiResult<String> {
-        return wrapExceptions("getSourceFile") {
+        return wrapRepoExceptions("getSourceFile") {
             bitbucketService.getRepositorySourceFile(workspaceSlug, repo, hash, path).toResult()
         }
     }
 
     override suspend fun createSnippet(title: String, filename: String, contents: String, private: Boolean): ApiResult<Unit> {
-        return wrapExceptions("createSnippet") {
+        return wrapRepoExceptions("createSnippet") {
             val body = MultipartBody.Part.createFormData("file", filename, RequestBody.create(MediaType.get("text/plain"), contents))
             bitbucketService.createSnippet(title, body, private).toEmptyResult()
         }
@@ -158,8 +158,8 @@ internal class BitbucketRepositoryImpl(
         _snippets.value = emptyList()
     }
 
-    /** Delegates to [wrapExceptions], passing in the class name here instead of requiring it of all callers */
-    private suspend fun <T : Any> wrapExceptions(methodName: String, block: suspend () -> ApiResult<T>): ApiResult<T> {
+    /** Delegates to [wrapRepoExceptions], passing in the class name here instead of requiring it of all callers */
+    private suspend fun <T : Any> wrapRepoExceptions(methodName: String, block: suspend () -> ApiResult<T>): ApiResult<T> {
         return wrapExceptions("BitbucketRepository", methodName, block)
     }
